requirements.txt
# Core dependencies
streamlit>=1.24.0
numpy>=1.24.0
pandas>=2.0.0
plotly>=5.14.0

# Vector store
faiss-cpu>=1.7.4

# LLM providers
openai>=1.0.0
cohere>=4.0.0
anthropic>=0.5.0

# CLI
click>=8.1.0
rich>=13.0.0

# Utilities
requests>=2.31.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
jinja2>=3.1.0

# ML/NLP
sentence-transformers>=2.5.1
torch>=2.2.0
tqdm>=4.66.2
scikit-learn>=1.4.0

# Document processing
beautifulsoup4>=4.12.3
markdown>=3.5.2
PyPDF2>=3.0.0

# Configuration
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Visualization
umap-learn>=0.5.5

# System monitoring
psutil>=5.9.0 